update project (implement new Codes)

+++ Common/AspNetCore/ClaimUtils.cs
@@ -4,7 +4,7 @@ namespace AspNetCore
 {
     public static class ClaimUtils
     {
-        public static int GetUserId(this ClaimsPrincipal principal)
+        public static long GetUserId(this ClaimsPrincipal principal)
         {
             if (principal == null)
                 throw new ArgumentNullException(nameof(principal));
